# [WIP] Speed up ONNX graphsurgeon toposort